+++ app/src/main/java/eu/kanade/tachiyomi/data/track/bangumi/BangumiModels.kt
@@ -16,7 +16,7 @@ data class Collection(
     val comment: String? = "",
     val ep_status: Int? = 0,
     val lasttouch: Int? = 0,
-    val rating: Float? = 0f,
+    val rating: Double? = 0.0,
     val status: Status? = Status(),
     val tag: List<String?>? = emptyList(),
     val user: User? = User(),
