+++ app/src/main/java/eu/kanade/tachiyomi/data/track/model/TrackSearch.kt
@@ -20,7 +20,7 @@ class TrackSearch : Track {
 
     override var total_chapters: Int = 0
 
-    override var score: Float = -1f
+    override var score: Double = -1.0
 
     override var status: Int = 0
